pyproject.toml
[project]
name = "ssg-hs-forensics-app"
version = "0.1.0"
description = "High school app for forensics"
authors = [
    {name = "John Leonard",email = "john@lowkeylabs.com"}
]
readme = "README.md"
homepage = "https://vcu-ssg.github.io/ssg-hs-forensics-app/"
repository = "https://github.com/vcu-ssg/ssg-hs-forensics-app"

packages = [
    { include = "ssg_hs_forensics_app", from = "src" }
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
click = "^8.1"
jupyter = ">=1.1.1,<2.0.0"
matplotlib = ">=3.10.5,<4.0.0"
plotly = ">=6.3.0,<7.0.0"
torchvision = "^0.23.0"
segment-anything = "^1.0"
ipykernel = "^6.30.1"
pip = "^25.2"
opencv-python = "^4.12.0.88"
pytest = "^9.0.1"
fastapi = "^0.121.3"
loguru = "^0.7.3"
tomli-w = "^1.2.0"

[tool.poetry.scripts]
sammy = "ssg_hs_forensics_app.cli._main:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
